chain/requirements.txt
fastapi~=0.116
uvicorn[standard]~=0.35
stellar-sdk~=13.0
python-dotenv>=1.0,<2
requests>=2.32,<3
bip-utils>=2.9,<3
